fix a big bug
Until last commit, SS and ESP does not contain appropriate value before jump to C language. I found it by using a function, in that function I made a pointer points to a local variable (set_gate() in kernel/traps.c), as I debugged with Bochs-debugger, the pointer can not get correct variable, at last I realised the stack was not set yet, SS.base is still the value before entering protected mode, so the bug is that, thanks God I found it. :)